Code cleanup and bookings query fix
